refactor into lib and example
