+++ examples/mario.rs
@@ -1,0 +1,30 @@
+use rasp_ws2812b::LedPanel;
+
+fn main() {
+    let mut panel = LedPanel::new(256);
+
+    let mario: Vec<&str> = [
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 1, 0, 0, 2, 2, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 2, 2,
+        2, 0, 1, 1, 1, 1, 3, 3, 2, 1, 1, 1, 1, 0, 1, 2, 2, 1, 3, 0, 3, 3, 1, 1, 1, 2, 2, 1, 1, 1,
+        3, 3, 3, 3, 1, 1, 0, 0, 3, 3, 3, 2, 3, 3, 3, 2, 2, 1, 2, 1, 3, 3, 3, 3, 2, 2, 2, 2, 2, 1,
+        1, 3, 3, 3, 3, 0, 0, 0, 0, 0, 0, 3, 3, 3, 3, 1, 1, 2, 2, 2, 2, 2, 3, 3, 0, 3, 1, 1, 2, 1,
+        2, 1, 3, 3, 2, 3, 3, 3, 0, 0, 1, 1, 3, 3, 3, 3, 1, 1, 0, 2, 1, 1, 1, 1, 3, 0, 0, 3, 0, 2,
+        2, 1, 2, 0, 1, 1, 1, 2, 3, 3, 1, 1, 1, 1, 0, 2, 2, 2, 1, 1, 0, 0, 1, 2, 2, 0, 0, 0, 0, 0,
+        0, 0, 2, 0, 0, 0, 0, 1, 2, 2, 2, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    ]
+    .iter()
+    .map(|val| {
+        match val {
+            0 => "000000", // Background
+            1 => "040101", // Boots
+            2 => "020100", // Skin
+            3 => "030000", // Clothes
+            _ => panic!("Invalid color"),
+        }
+    })
+    .collect();
+
+    panel.convert_and_write(mario.as_slice());
+}
