--- src/mario
@@ -1,24 +1,0 @@
-// 0 = WHITE
-// 1 = BROWN
-// 2 = SKIN
-// 3 = CLOTHES
-
-
-[
-  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-  1, 0, 0, 2, 2, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-  0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 2, 2, 2, 0, 1, 1,
-  1, 1, 3, 3, 2, 1, 1, 1, 1, 0, 1, 2, 2, 1, 3, 0,
-  3, 3, 1, 1, 1, 2, 2, 1, 1, 1, 3, 3, 3, 3, 1, 1,
-  0, 0, 3, 3, 3, 2, 3, 3, 3, 2, 2, 1, 2, 1, 3, 3,
-  3, 3, 2, 2, 2, 2, 2, 1, 1, 3, 3, 3, 3, 0, 0, 0,
-  0, 0, 0, 3, 3, 3, 3, 1, 1, 2, 2, 2, 2, 2, 3, 3,
-  0, 3, 1, 1, 2, 1, 2, 1, 3, 3, 2, 3, 3, 3, 0, 0,
-  1, 1, 3, 3, 3, 3, 1, 1, 0, 2, 1, 1, 1, 1, 3, 0,
-  0, 3, 0, 2, 2, 1, 2, 0, 1, 1, 1, 2, 3, 3, 1, 1,
-  1, 1, 0, 2, 2, 2, 1, 1, 0, 0, 1, 2, 2, 0, 0, 0,
-  0, 0, 0, 0, 2, 0, 0, 0, 0, 1, 2, 2, 2, 0, 0, 1,
-  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-]
